+++ docs/history.md
@@ -14,6 +14,11 @@ V1.0.8
 *New modules*
 
 -   panos\_management\_profile: Manages interface management profiles
+-   panos\_ike\_crypto\_profile: Use the IKE Crypto Profiles page to specify protocols and algorithms for 
+identification, authentication, and encryption (IKEv1 or IKEv2, Phase 1).
+-   panos\_ipsec\_profile: Configures IPSec Crypto profile on the firewall with subset of settings.
+-   panos\_ike\_gateway: Configures IKE gateway on the firewall with subset of settings.
+-   panos\_ipsec\_tunnel: Configure data-port (DP) network interface for DHCP. By default DP interfaces are static.
 
 *Removed modules*